pyproject.toml
[project]
name = "wolkenlose-ki-fuer-zu-hause"
version = "1.0.0"
description = "Material used for talk 'Wolkenlose KI für zu Hause' at the Linuxtage in Graz on 2025-04-26."
requires-python = ">=3.12"
dependencies = [
    "marker-pdf[full]>=1.6.2",
    "notebook>=7.4.0",
    "pandas>=2.2.3",
    "psycopg[binary]>=3.2.6",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
